Solved issue of yer-less words.
Next step is moving out of the scope of individual inflected words, to only determine if there is a yer somewhere in the paradigm. If not, *then* add to the yer_not_found list. Otherwise, if a yer is found but the other forms lack a yer, mark the paradigm as yer-bearing but note each form for having a yer or not.